+++ cases/pfd_4_1_1.json
@@ -1,0 +1,1 @@
+{"basemva": 100.0, "ws": 376.99111843077515, "bus_num": [1, 2, 3, 4], "bus_type": [3, 1, 1, 2], "bus_Vm": [1.04, 0.98071, 0.9395, 0.94], "bus_Va": [0.0, -0.11096105252479148, -0.1906510408416006, -0.1835737307247636], "bus_kV": [24.959999084472656, 225.5643768310547, 216.08587646484375, 22.559999465942383], "bus_basekV": [24.0, 230.0, 230.0, 24.0], "bus_name": ["BUS 1       ", "BUS 2       ", "BUS 3       ", "BUS 4       "], "load_id": ["1 ", "1 "], "load_bus": [2, 3], "load_Z": ["-0j", "-0j"], "load_I": ["0j", "0j"], "load_P": ["(105.04800415039062+58.23600387573242j)", "(100+40j)"], "load_MW": [105.048, 100.0], "load_Mvar": [58.236, 40.0], "ibr_bus": [], "ibr_id": [], "ibr_MW": [], "ibr_Mvar": [], "ibr_MVA_base": [], "gen_id": ["1 ", "1 "], "gen_bus": [1, 4], "gen_S": ["(196.0782470703125+117.93246459960938j)", "(10+0.7816917896270752j)"], "gen_mod": [0, 0], "gen_MW": [196.078, 10.0], "gen_Mvar": [117.932, 0.782], "gen_MVA_base": [555.0, 555.0], "gen_status": [1, 1], "line_from": [2], "line_to": [3], "line_id": ["1 "], "line_P": [91.03024291992188], "line_Q": [31.815231323242188], "line_RX": ["(0.01+0.08499999999999999j)"], "line_chg": [0.176], "xfmr_from": [2, 3], "xfmr_to": [1, 4], "xfmr_id": ["1 ", "1 "], "xfmr_P": [196.0782470703125, -10.0], "xfmr_Q": [117.93246459960938, -0.7105262279510498], "xfmr_RX": ["0.0576j", "0.0625j"], "xfmr_k": [1.0, 1.0], "shnt_bus": [], "shnt_id": [], "shnt_gb": [], "shnt_sw_bus": [], "shnt_sw_gb": []}
