+++ main_plot_2area.m
@@ -7,7 +7,7 @@ dataTld = readtable('emt_x_load.csv');
 dataTbus = readtable('emt_x_bus.csv');
 dataT_voltage = readtable('emt_3phaseV.csv');
 
-tend = 10;
+tend = 20;
 dr = 10;
 ts = dr*50e-6;
